+++ src/WinPaperWalls/Services/IDesktopWallpaperService.cs
@@ -1,9 +1,0 @@
-using WinPaperWalls.Interop;
-
-namespace WinPaperWalls.Services;
-
-public interface IDesktopWallpaperService
-{
-	void SetWallpaper(string filePath, WallpaperStyle style);
-	string? GetCurrentWallpaperPath();
-}