+++ src/WinPaperWalls/Services/ILogBundleService.cs
@@ -1,6 +1,0 @@
-namespace WinPaperWalls.Services;
-
-internal interface ILogBundleService
-{
-	Task<string> CreateBugReportAsync();
-}